test-requirements.txt
pyyaml>=6.0
testfixtures>=7.0
deepdiff>=5.6.0
requests
msal
pytest==6.2.5
pytest-cov==3.0.0
pytablewriter
